requirements.txt
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Django==2.2.4
django-cors-headers==2.4.0
django-templated-mail==1.1.1
django-trench==0.2.2
django-username-email==2.2.4
djangorestframework==3.9.1
djoser==1.7.0
ecdsa==0.14.1
future==0.18.2
gunicorn==19.6.0
idna==2.8
mysqlclient==1.4.2
pyasn1==0.4.8
pycparser==2.19
Pygments==2.2.0
PyJWT==1.7.1
pyotp==2.2.7
PySocks==1.7.0
python-jose==3.0.1
python-memcached==1.59
pytz==2018.5
requests==2.22.0
rsa==4.0
six==1.12.0
smsapi-client==2.3.0
sqlparse==0.3.0
twilio==6.27.1
urllib3==1.25.3
yubico-client==1.10.0
